refactor: Tag-related implementations
* Fix stateful use of IGitModule
* Fix abstraction level removing UI-related calls from GitTagController
* Split and rework unit tests
