+++ GitCommands/Git/Tag/GitTagController.cs
@@ -1,9 +1,8 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using System.IO.Abstractions;
-using System.Windows.Forms;
-using GitUIPluginInterfaces;
 
-namespace GitCommands.Git
+namespace GitCommands.Git.Tag
 {
     public interface IGitTagController
     {
@@ -11,42 +10,39 @@ namespace GitCommands.Git
         /// Create the Tag depending on input parameter.
         /// </summary>
         /// <returns><see langword="true"/> if create tag command succeeded, <see langword="false"/> otherwise</returns>
-        bool CreateTag(GitCreateTagArgs args, IWin32Window parentForm);
+        GitCreateTagCmd GetCreateTagCommand(GitCreateTagArgs args);
     }
-
 
     public class GitTagController : IGitTagController
     {
-        private readonly IGitUICommands _uiCommands;
-        private readonly IGitModule _module;
+        private readonly Func<string> _getWorkingDir;
         private readonly IFileSystem _fileSystem;
 
-        public GitTagController(IGitUICommands uiCommands, IGitModule module, IFileSystem fileSystem)
+        public GitTagController(Func<string> getWorkingDir, IFileSystem fileSystem)
         {
-            _module = module;
-            _uiCommands = uiCommands;
+            _getWorkingDir = getWorkingDir;
             _fileSystem = fileSystem;
         }
 
-        public GitTagController(IGitUICommands uiCommands, IGitModule module)
-            : this(uiCommands, module, new FileSystem())
+        public GitTagController(Func<string> getWorkingDir)
+            : this(getWorkingDir, new FileSystem())
         { }
 
         /// <summary>
         /// Create the Tag depending on input parameter.
         /// </summary>
         /// <returns>Output string from RunGitCmd.</returns>
-        public bool CreateTag(GitCreateTagArgs args, IWin32Window parentForm)
+        public GitCreateTagCmd GetCreateTagCommand(GitCreateTagArgs args)
         {
             string tagMessageFileName = null;
             if (args.Operation.CanProvideMessage())
             {
-                tagMessageFileName = Path.Combine(_module.WorkingDirGitDir, "TAGMESSAGE");
+                tagMessageFileName = Path.Combine(_getWorkingDir(), "TAGMESSAGE");
                 _fileSystem.File.WriteAllText(tagMessageFileName, args.TagMessage);
             }
 
             var createTagCmd = new GitCreateTagCmd(args, tagMessageFileName);
-            return _uiCommands.StartCommandLineProcessDialog(createTagCmd, parentForm);
+            return createTagCmd;
         }
     }
 }
