+++ GitCommands/Git/Tag/TagOperation.cs
@@ -1,0 +1,10 @@
+namespace GitCommands.Git.Tag
+{
+    public enum TagOperation
+    {
+        Lightweight = 0,
+        Annotate,
+        SignWithDefaultKey,
+        SignWithSpecificKey
+    };
+}
