+++ UnitTests/GitCommandsTests/GitCommandsTests.csproj
@@ -73,6 +73,7 @@
     <Compile Include="GitRevisionInfoProviderTests.cs" />
     <Compile Include="Git\EncodingHelperTest.cs" />
     <Compile Include="Git\GitBlameHeaderTest.cs" />
+    <Compile Include="Git\Tag\GitCreateTagCmdTests.cs" />
     <Compile Include="Git\RevisionDiffProviderTest.cs" />
     <Compile Include="Git\GitBlameTest.cs" />
     <Compile Include="Git\GitBranchNameNormaliserTest.cs" />
@@ -82,7 +83,7 @@
     <Compile Include="Git\GitCommandHelpersTest.cs" />
     <Compile Include="Git\IndexLockManagerTests.cs" />
     <Compile Include="Git\GitDirectoryResolverTests.cs" />
-    <Compile Include="Git\GitTagControllerTest.cs" />
+    <Compile Include="Git\Tag\GitTagControllerTest.cs" />
     <Compile Include="Git\GitTreeParserTests.cs" />
     <Compile Include="Helpers\FileInfoExtensions.cs" />
     <Compile Include="Helpers\FileInfoExtensionsTest.cs" />
